🕸️ GraphQL schema + resolvers plugged
Untested for the time being, but works as proof of concept. I want to
plug in comments on supported entities so that the funcitonality can be
left latent until frontend is ready for it.

This should speed the post branch a little, as the system is
plug-and-play. I am optimistic about the upcoming merge.