Starting to get tick code in order
This code needs testing, and I'm not sure if we have a plan for GQL
layer testing. It seems as if there is a bit of spaghetti insofar as
auth is concerned with our backend.

This branch will also need an accompanying branch on open-tacos, as the
changes here are essentially entirely breaking (Though there is a chance
that supplying unneeded params will simply pass over, as we aren't using
ZOD or anything like that)