Migration to ApolloServer V4
Apollo Server 3 is officially deprecated, with end-of-life scheduled for
22 October 2023.

If you skim through apollographql.com/docs/apollo-server/migration this
is pretty much the result you get.

With some type repair in middleware.ts I can be a little more confident,
but since we have no integration testing for the GQL layer it's hard to
know if the API is still producing the same Data.