refactor: remove .js ext from ALL local imports across the codebase
ideally, imports should reference the source file, not the build file.
the triger for this change was a requirement for jest to work directly with typescript files rather than the build output. but not just that, this change should move the project to a more standard direction and also improve code editors to better understand the code structure since now they reference the source files.