Delete Area Transformer
My confidence in this deletion is a little lower than it initially was.
I remember checking the code archeology the first time around and
thinking that we must have been done with importing data (especially
since descriptions have become mutable and I had assumed importing
might result in unintensional overwrites.)

However this change to the code indicates otherwise. Could someone
indicate where in the workflow this code is still in use? If it's still
in use then obviously I actually do need to fix this up for the PR to
make any sense...

Happy new year, fellows!