Add support for differnet GCS adapters
I have chosen for the time being to support either the PUSH or PULL
based subscribers, and for now I haven't done anything clever with
deletion events though realistically we could do that to make sure media
gets properly removed - though it would require some authentication
hoops to be jumped through.

Ideally I'd like for us to choose one of these that we would like to
support and then leave the other at the wayside, opting for a mock
service to be used with all other development environments.