Rather than keep media pending in the open, hide it
We leave it to the client to recall its own promised media, because the
alternative involves convuluting the GQL schema and potentially
poisoning the consistency.

In the event that promised media gets dropped, so what? the server will
statelessly deal with the timeout by quietly terminating the pending
media.

This means we need a test case for re-submitting an idential pending
media promise. In this case I imagine we can yield back the existing
pending object - unless the expiry is imminent in which case we can
re-sign a url and send it back.