Add tests and fixes for issues revealed by tests
As far as I can tell these are quite comprehensive, I need to run a
realworld test of the google pull subscriber load balancing so that I
can see if the complexity is needless. That will be the next commit.

This commit contains working code that should be safe to integrate into
the existing services quite easily so long as the infrastructure has the
notifications configured along with it.

I've tried to make the tests catch any funny business, so it should be
safe for any developer to come and poke around without disrupting
anything. The integration tests are the end-all since if they pass, the
code should be totally funcitonal.