Pull in new vitest version (3) + refactor some test infra
Shaving off some additional seconds and some extra logging noise, this
is now at a point where I would like it to be. The tests can be run
pretty much continuously through the development cycle and can be used
to track down all manner of gremlin

TODO: The history behavior is still a mystery to me. I need to know some
abstract about it so I can understand how to test it a bit better but I
have been avoiding looking closely at it