Assert json types is long since deprecated
I'm getting tired of seeing the warnings, so they are outa here. I'm
running nodejs 18 for development on this project, but moving forward to
newer versions should not present an issue.

I did a cursory check with nodejs 20, and 23, and it did not seem like
the import for json had any problems - though the addition of the `with`
keyword was made in some recent versions and I am not sure of the impact
this has on the behavior of the import.