fix for #468 and partial solution to #413
I've added some e2e tests to make sure things are above board but there
is a bit of potential uncertainty. Frustratingly, this is a part of the
GQL architechture that leads to a growing number of utility types that
developers need to keep track of.

Because I do not have a good insight into how many services still depend
on the old user profile resolutions I have left them untouched - though
it could be argued they need to be deprecated.

As far as resolving user nodes on Media, this is simply a danger we are
going to accept. We'd like to PREVENT resolving greedily in places where
it's not necessary but really there's nothing we can do to prevent
fronent devs from doing silly things