Support node v22
This is the current LTS, so we NEED to get here, but upgrading gives us
a couple of problems:

1. ts-standard doesn't jive well with modern typescript, and the mystery
   of import assertions has only deepened with age.
2. I have therefore moved to the non-deprecated versions of these tools
   (which leaves us with a tasty >2k lint issues)

resolves #473

There are now two choices: Disable the failing rules to get rid of the
two thousand lint issues, or simply fix them so we can go forward with
modern eslint standards. The rules do appear to relate to newer TS
practices, rather than noise caused by adding semicolons or changing
every instance of a single quote to a double quote. I will make a
seperate branch so we can look at the fixes as opposed to just silencing
them